refactor axios call
