+++ src/api/axiosCall.ts
@@ -1,9 +1,9 @@
 import axios from 'axios';
 
-export const getCountryByCode = async (url: string) => {  
+export const axiosCall = async (url: string) => {  
   try {
     const response = await axios.get(url);
-    return response.data[0];
+    return response.data;
   } catch (error) {
     throw error;
   }
