+++ src/pages/index.tsx
@@ -1,16 +1,28 @@
-import type { GetServerSideProps } from 'next';
+// components
+import MainLayout from '@/components/layouts/mainLayout/mainLayout';
+import CountryCards from '@/components/common/countryCards/countryCards';
+// consts
+import { ENDPOINT_GET_ALL_COUNTRIES } from '@/consts/endpoints';
 
-const Index = () => {
-  return null;
-};
+import { axiosCall } from '@/api/axiosCall';
 
-export const getServerSideProps: GetServerSideProps = async () => {
-  return {
-    redirect: {
-      destination: '/home',
-      permanent: false,
-    },
-  };
-};
+const Home = () => {
+  return (
+    <MainLayout {...{ fetcher: axiosCall, url: ENDPOINT_GET_ALL_COUNTRIES }}>      
+      {({ data, error, isLoading, isValidating }) => {
+        return (
+          <CountryCards 
+            {...{
+              data,
+              error,
+              isValidating,
+              isLoading,
+            }}
+          />
+        );
+      }}
+    </MainLayout>
+  );
+}
 
-export default Index;
+export default Home;
